Rewrite for JUCE 6

+++ Source/PluginProcessor.h
@@ -15,29 +15,28 @@
 #include <ATK/Core/InPointerFilter.h>
 #include <ATK/Core/OutPointerFilter.h>
 
-#include <ATK/Tools/OversamplingFilter.h>
 #include <ATK/Tools/DecimationFilter.h>
 #include <ATK/Tools/DryWetFilter.h>
+#include <ATK/Tools/OversamplingFilter.h>
 #include <ATK/Tools/VolumeFilter.h>
 
 #include <ATK/EQ/ButterworthFilter.h>
 #include <ATK/EQ/IIRFilter.h>
 #include <ATK/EQ/ToneStackFilter.h>
 
-#include <ATK/Preamplifier/Triode2Filter.h>
 #include <ATK/Preamplifier/DempwolfTriodeFunction.h>
-
+#include <ATK/Preamplifier/Triode2Filter.h>
 
 //==============================================================================
 /**
-*/
+ */
 static const double minVolume = -40;
 static const double maxVolume = 40;
 static const double minGain = -40;
 static const double originGain = -24;
 static const double maxGain = 40;
 
-class ATKBassPreampAudioProcessor  : public AudioProcessor
+class ATKBassPreampAudioProcessor: public AudioProcessor
 {
 public:
   //==============================================================================
@@ -45,14 +44,14 @@ public:
   ~ATKBassPreampAudioProcessor();
 
   //==============================================================================
-  void prepareToPlay (double sampleRate, int samplesPerBlock) override;
+  void prepareToPlay(double sampleRate, int samplesPerBlock) override;
   void releaseResources() override;
 
 #ifndef JucePlugin_PreferredChannelConfigurations
-  bool isBusesLayoutSupported (const BusesLayout& layouts) const override;
+  bool isBusesLayoutSupported(const BusesLayout& layouts) const override;
 #endif
 
-  void processBlock (AudioSampleBuffer&, MidiBuffer&) override;
+  void processBlock(AudioSampleBuffer&, MidiBuffer&) override;
 
   //==============================================================================
   AudioProcessorEditor* createEditor() override;
@@ -68,31 +67,31 @@ public:
   //==============================================================================
   int getNumPrograms() override;
   int getCurrentProgram() override;
-  void setCurrentProgram (int index) override;
-  const String getProgramName (int index) override;
-  void changeProgramName (int index, const String& newName) override;
+  void setCurrentProgram(int index) override;
+  const String getProgramName(int index) override;
+  void changeProgramName(int index, const String& newName) override;
 
   //==============================================================================
-  void getStateInformation (MemoryBlock& destData) override;
-  void setStateInformation (const void* data, int sizeInBytes) override;
-  
+  void getStateInformation(MemoryBlock& destData) override;
+  void setStateInformation(const void* data, int sizeInBytes) override;
+
 private:
   //==============================================================================
-  JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (ATKBassPreampAudioProcessor)
-    
+  JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(ATKBassPreampAudioProcessor)
+
   ATK::InPointerFilter<float> inFilter;
-  ATK::VolumeFilter<float> levelFilter;
-  ATK::OversamplingFilter<float, ATK::Oversampling6points5order_4<float> > oversamplingFilter;
-  ATK::Triode2Filter<float, ATK::DempwolfTriodeFunction<float>> overdriveFilter;
-  ATK::IIRFilter<ATK::ButterworthLowPassCoefficients<float> > lowpassFilter;
-  ATK::DecimationFilter<float> decimationFilter;
-  ATK::IIRFilter<ATK::ToneStackCoefficients<float> > toneFilter;
-  ATK::VolumeFilter<float> volumeFilter;
-  ATK::DryWetFilter<float> dryWetFilter;
+  ATK::VolumeFilter<double> levelFilter;
+  ATK::OversamplingFilter<double, ATK::Oversampling6points5order_4<double>> oversamplingFilter;
+  ATK::Triode2Filter<double, ATK::DempwolfTriodeFunction<double>> overdriveFilter;
+  ATK::IIRFilter<ATK::ButterworthLowPassCoefficients<double>> lowpassFilter;
+  ATK::DecimationFilter<double> decimationFilter;
+  ATK::IIRFilter<ATK::ToneStackCoefficients<double>> toneFilter;
+  ATK::VolumeFilter<double> volumeFilter;
+  ATK::DryWetFilter<double> dryWetFilter;
   ATK::OutPointerFilter<float> outFilter;
 
   AudioProcessorValueTreeState parameters;
-  
+
   float old_gain;
   float old_bass;
   float old_medium;
